Refactor: changed FeedItemRoot and FeedImageModel location to an extension in RemoteFeedLoader.swift in order to set those classes as private since they are not used outside this module
